Feat(client/style): Dashboard styled
