--- src/components/DashBoard/dashboard.css
@@ -1,3 +1,0 @@
-.dashBoard__container{
-    min-height: 100vh;
-}
